Fixed a few warnings.
Rev: src/modules/HTTPLoop/accept_and_parse.c:1.23
Rev: src/modules/HTTPLoop/accept_and_parse.h:1.8
Rev: src/modules/HTTPLoop/requestobject.c:1.18
Rev: src/modules/Image/dct.c:1.18
Rev: src/modules/Image/font.c:1.67
Rev: src/modules/Image/matrix.c:1.34

+++ src/modules/HTTPLoop/accept_and_parse.h
@@ -83,7 +83,7 @@ struct cache
   struct cache_entry *htable[CACHE_HTABLE_SIZE];
   unsigned LONG_LONG size, entries, max_size;
   unsigned LONG_LONG hits, misses, stale;
-  unsigned int num_requests, sent_data, received_data;
+  size_t num_requests, sent_data, received_data;
   int gone;
 };
